Add search to include skills and client
